fix: check if click event is user input or programmatic
Since the extension is using programmatic clicks, and ALSO saves clicks which override the user selected values, a distinction needs to be made between user input and the extension's own clicks. Only user input should be used as the override. This is only relevant in playlists where the override value survives between videos.